refactor: move unit tests to its own place
